requirements/release.txt
bump2version
check-manifest
crowdin-api-client
click
jupyterlab-translate
pip
pre-commit
pyyaml
requests
setuptools
twine
wheel
